DEBT: Added CI badge and cleaned up various documentation errors and omissions (#21)
- **DEBT:** Added CI badge
- **DEBT:** ⚡ Performance Enhancements:
  - Unified `allocs/op`: Reduced allocations per operation to 1 for both ASCII and Unicode alphabets regardless of ID length, enhancing memory efficiency across all ID generations.
  - Decreased `ns/op`: Further optimized the Nano ID generation process to lower nanoseconds per operation, resulting in faster ID creation.

+++ nanoid_benchmark_test.go
@@ -171,7 +171,7 @@ func BenchmarkNanoIDGeneration(b *testing.B) {
 			// Initialize the generator without passing 'nil'
 			gen, err := NewGenerator(
 				WithAlphabet(alphabet),
-				WithLengthHint(int(mean)),
+				WithLengthHint(uint16(mean)),
 			)
 			if err != nil {
 				b.Fatalf("Failed to create generator with %s alphabet of length %d: %v", alphabetType, alphaLen, err)
@@ -224,7 +224,7 @@ func BenchmarkNanoIDGenerationParallel(b *testing.B) {
 			// Initialize the generator without passing 'nil'
 			gen, err := NewGenerator(
 				WithAlphabet(alphabet),
-				WithLengthHint(int(mean)),
+				WithLengthHint(uint16(mean)),
 			)
 			if err != nil {
 				b.Fatalf("Failed to create generator with %s alphabet of length %d: %v", alphabetType, alphaLen, err)
@@ -279,7 +279,7 @@ func BenchmarkNanoIDWithVaryingAlphabetLengths(b *testing.B) {
 			// Initialize the generator without passing 'nil'
 			gen, err := NewGenerator(
 				WithAlphabet(alphabet),
-				WithLengthHint(int(mean)),
+				WithLengthHint(uint16(mean)),
 			)
 			if err != nil {
 				b.Fatalf("Failed to create generator with %s alphabet of length %d: %v", alphabetType, alphaLen, err)
